--- ws/src/main/java/com/hoaxify/ws/configuration/CorsConfig.java
@@ -1,34 +1,0 @@
-package com.hoaxify.ws.configuration;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-import org.springframework.web.cors.CorsConfigurationSource;
-import java.util.List;
-
-@Configuration
-public class CorsConfig {
-
-    @Bean
-    public CorsConfigurationSource corsConfigurationSource() {
-        CorsConfiguration configuration = new CorsConfiguration();
-
-        configuration.setAllowedOrigins(List.of(
-            "http://localhost:3000",
-            "https://ws-spring-app.vercel.app",
-            "https://ws-spring-app-git-uat-erengthbs-projects.vercel.app"
-        ));
-        configuration.setAllowedMethods(List.of("GET", "POST", "PUT", "DELETE", "OPTIONS"));
-        configuration.setAllowedHeaders(List.of("*"));
-        configuration.setAllowCredentials(true);
-
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-
-        // Tüm path'ler için bu CORS ayarları geçerli
-        source.registerCorsConfiguration("/**", configuration);
-
-        return source;
-    }
-}
-
