+++ ws/src/main/resources/application.yaml
@@ -56,6 +56,9 @@ spring:
       on-profile: uat
   datasource:
     driver-class-name: org.postgresql.Driver
+    url: ${DATABASE_URL}
+    username: ${POSTGRES_USER}
+    password: ${POSTGRES_PASSWORD}
   jpa:
     hibernate:
       ddl-auto: update
@@ -74,6 +77,9 @@ spring:
       on-profile: production
   datasource:
     driver-class-name: org.postgresql.Driver
+    url: ${DATABASE_URL}
+    username: ${POSTGRES_USER}
+    password: ${POSTGRES_PASSWORD}
   jpa:
     hibernate:
       ddl-auto: update
